# api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemoryRecord": {
        "description": "Durable event-memory contract with evidence and temporal provenance.",
        "properties": {
          "confidence": {
            "default": 1.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "content": {
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "derivation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Derivation"
          },
          "derived_from": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Derived From",
            "type": "array"
          },
          "embedding_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Version"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/EntityRef"
            },
            "title": "Entities",
            "type": "array"
          },
          "event_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Time"
          },
          "evidence_refs": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRef"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "memory_id": {
            "format": "uuid",
            "title": "Memory Id",
            "type": "string"
          },
          "memory_kind": {
            "$ref": "#/components/schemas/MemoryKind",
            "default": "fact"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "normalized_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Content"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/RelationRef"
            },
            "title": "Relations",
            "type": "array"
          },
          "roles": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Roles",
            "type": "object"
          },
          "schema_version": {
            "const": "memory.v1",
            "default": "memory.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "status": {
            "$ref": "#/components/schemas/MemoryStatus",
            "default": "active"
          },
          "superseded_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Superseded By"
          },
          "supersedes": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Supersedes",
            "type": "array"
          },
          "synthetic": {
            "default": false,
            "title": "Synthetic",
            "type": "boolean"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "minLength": 1,
            "title": "User Id",
            "type": "string"
          },
          "utility": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Utility",
            "type": "number"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          }
        },
        "required": [
          "user_id",
          "content"
        ],
        "title": "MemoryRecord",
        "type": "object"
      },
...
      "_V1ExplainResponse": {
        "properties": {
          "memory": {
            "$ref": "#/components/schemas/_V1MemoryResponse"
          },
          "related": {
            "items": {
              "$ref": "#/components/schemas/_V1MemoryResponse"
            },
            "title": "Related",
            "type": "array"
          }
        },
        "required": [
          "memory",
          "related"
        ],
        "title": "_V1ExplainResponse",
        "type": "object"
      },
      "_V1FeedbackRequest": {
        "properties": {
          "outcome": {
            "minLength": 1,
            "title": "Outcome",
            "type": "string"
          },
          "rating": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          }
        },
        "required": [
          "outcome"
        ],
        "title": "_V1FeedbackRequest",
        "type": "object"
      },
...
      "_V1MemorySearchHitResponse": {
        "properties": {
          "memory": {
            "$ref": "#/components/schemas/_V1MemoryResponse"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "memory",
          "score",
          "reason"
        ],
        "title": "_V1MemorySearchHitResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "EvoEventMem",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/memories": {
      "post": {
        "operationId": "write_memory_v1_memories_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryRecord"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_V1MemoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write Memory"
      }
    },
    "/v1/memories/search": {
      "get": {
        "operationId": "search_memories_v1_memories_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/_V1MemorySearchHitResponse"
                  },
                  "title": "Response Search Memories V1 Memories Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Memories"
      }
    },
    "/v1/memories/{memory_id}/explain": {
      "get": {
        "operationId": "explain_memory_v1_memories__memory_id__explain_get",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_V1ExplainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Explain Memory"
      }
    },
    "/v1/memories/{memory_id}/feedback": {
      "post": {
        "operationId": "feedback_memory_v1_memories__memory_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_V1FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_V1MemoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feedback Memory"
      }
    },
    "/v1/memories/{memory_id}/forget": {
      "post": {
        "operationId": "forget_memory_v1_memories__memory_id__forget_post",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-Tenant-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Session-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_V1MemoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forget Memory"
      }
    }
  }
}